+++ src/includes/invoice.h
@@ -4,4 +4,4 @@
 #include <vector>
 
 std::string currentDateTime();
-void generateInvoice(std::map<std::string, float> &details);
+void generateInvoice(std::map<std::string, double> &details);
